Move initialization of signals up into the parent class
